Updated the GameDat.exe
Made it so the game can display up to 4 lines of script instead of 3, greatly increasing the max line length. Made the text a bit smaller and crunched together, but it'll have to do until I can figure out how to add commands in the script to add more lines, or until I can figure out how to supress error messages.